add test, deploy to rinkeby

+++ contracts/interfaces/IExampleOracleSimple.sol
@@ -2,5 +2,8 @@
 pragma solidity 0.8.13;
 
 interface IExampleOracleSimple {
-    function consult(address token, uint amountIn) external view returns (uint amountOut);
+    function consult(address token, uint256 amountIn)
+        external
+        view
+        returns (uint256 amountOut);
 }
